requirement.txt
django
djangorestframework
psycopg2
psycopg2-binary
django-cors-headers